Reorganize entire lib
