+++ ftc-visionlib-java/src/main/java/org/lasarobotics/vision/android/Cameras.java
@@ -1,7 +1,9 @@
-package org.lasarobotics.vision;
+package org.lasarobotics.vision.android;
 
 import android.content.Context;
 import android.content.pm.PackageManager;
+
+import org.lasarobotics.vision.android.Camera;
 
 /**
  * Implements the Android camera
