+++ ftc-visionlib-java/src/main/java/org/lasarobotics/vision/android/Util.java
@@ -1,4 +1,4 @@
-package org.lasarobotics.vision;
+package org.lasarobotics.vision.android;
 
 import android.os.Environment;
