+++ ftc-visionlib-java/src/main/java/org/lasarobotics/vision/detection/BlobDetection.java
@@ -1,0 +1,8 @@
+package org.lasarobotics.vision.detection;
+
+/**
+ * Implements detection based on color or similar data
+ */
+public class BlobDetection {
+
+}
