+++ ftc-visionlib-java/src/main/java/org/lasarobotics/vision/detection/EdgeDetection.java
@@ -1,0 +1,8 @@
+package org.lasarobotics.vision.detection;
+
+/**
+ * Implements object recognition based on edge detection methods
+ */
+public class EdgeDetection {
+
+}
