+++ ftc-visionlib-java/src/main/java/org/lasarobotics/vision/util/FPS.java
@@ -1,4 +1,4 @@
-package org.lasarobotics.vision;
+package org.lasarobotics.vision.util;
 
 import org.lasarobotics.vision.util.RollingAverage;
